Head collision update, weight plugin
Updated the collision of the head to more closely resemble the correct dimensions. Created a plugin to enable adjustable weights to be added to the mouth and tail of the robot, simulating loading or potential tail weights. Hopefully this will allow the trained controller to be able to rapidly respond to changes in the environment