Training motion controller in progress.
This is quite a large commit, I've added a custom Gazebo contact detection plugin, custom FREA messages package, and sorted out Melodic->Noetic changes. Lots of little fixes to scripts and files that were untested in earlier commits. Training seems to work for a bit before gazebo freezes (without crashing, so I don't seem to be able to gdb the damn thing :(, but we'll figure it out eventually).